Fix and warn about implicit sign conversions in ebpf code
The sign extension generates bad code, so better to avoid it
if unneeded. Fix code where possible, and enable warnings to
be more explicit of this.

Convert also for loop main condition to be on unsigned when
possible.

Reduces ebpf blob size a bit, generates faster code and
is more pedantic about code correctness.

+++ support/ebpf/native_stack_trace.ebpf.c
@@ -150,7 +150,7 @@ static EBPF_INLINE int unwind_native(struct pt_regs *ctx)
   Trace *trace = &record->trace;
   int unwinder;
   ErrorCode error;
-  for (int i = 0; i < NATIVE_FRAMES_PER_PROGRAM; i++) {
+  for (u64 i = 0; i < NATIVE_FRAMES_PER_PROGRAM; i++) {
     unwinder = PROG_UNWIND_STOP;
 
     // Unwind native code
